requirements.txt
# Core TensorFlow and ML dependencies
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
keras==3.9.1
numpy>=1.26.0,<2.0.0
protobuf==5.29.4
absl-py==2.2.1
astunparse==1.6.3
google-pasta==0.2.0
six==1.17.0
typing-extensions==4.12.2
opt-einsum==3.4.0
gast==0.6.0
flatbuffers==25.2.10
libclang==18.1.1
requests==2.32.3
h5py==3.13.0
wrapt==1.17.2
termcolor==3.0.1
setuptools==59.6.0
tensorboard==2.19.0
ml-dtypes==0.5.1
grpcio==1.71.0
packaging==24.2

# Visualization + image processing
matplotlib==3.10.6
pillow==11.3.0
opencv-python>=4.8.0,<5.0.0

# Kaggle dataset utilities
kaggle==1.7.4.5
python-slugify==8.0.4
text-unidecode==1.3
tqdm==4.67.1
bleach==6.2.0
webencodings==0.5.1

# FastAPI and web server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Environment and utilities
python-dotenv==1.0.0

# Database and AI services
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai==1.12.0

# Document processing
docx2txt==0.8
langchain==0.2.16